pyproject.toml
[project]
name = "remote-mcp-azure-function"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.23.0",
    "httpx>=0.28.1",
]
